rename ray dir to runtime
