+++ wasm/src/runtime/core/global_state_accessor.rs
